requirements.txt
confuse==2.0.1
DiscordLevelingCard==0.6.0
tinydb==4.8.0